ColourGenerator/backend/requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
numpy==2.3.3
pillow==11.3.0
scikit-learn==1.7.2
scipy==1.16.2
threadpoolctl==3.6.0
Werkzeug==3.1.3
